Remove test storage
