+++ irohad/ametsuchi/CMakeLists.txt
@@ -5,8 +5,6 @@ add_library(ametsuchi
     impl/temporary_wsv_impl.cpp
     impl/mutable_storage_impl.cpp
 
-    impl/test_storage_impl.cpp
-
     impl/postgres_wsv_query.cpp
     impl/postgres_wsv_command.cpp
     impl/peer_query_wsv.cpp
