+++ test/module/irohad/ametsuchi/CMakeLists.txt
@@ -16,13 +16,6 @@ target_link_libraries(flat_file_test
     libs_common
     )
 
-addtest(testing_storage_test testing_storage_test.cpp)
-target_link_libraries(testing_storage_test
-    ametsuchi
-    test_block_generator
-    libs_common
-    )
-
 add_library(ametsuchi_fixture INTERFACE)
 target_link_libraries(ametsuchi_fixture INTERFACE
     pqxx
